requirements.txt
Jinja2==3.0.3
python_frontmatter==1.0.0
PyYAML==6.0
requests==2.27.1
